requirements.txt
Flask==2.3.0
tensorflow==2.13.0
Pillow==9.5.0
numpy==1.24.3
gunicorn==20.1.0
